Organize Value types to their own Type namespace
