+++ Assets/Scripts/Interpreter/Cmd/Log.cs
@@ -1,3 +1,4 @@
+using Interpreter.Eval.Type;
 using UnityEngine;
 
 namespace Interpreter.Cmd
@@ -5,7 +6,7 @@ namespace Interpreter.Cmd
     // [UserFunction("console_log")]
     public class Log
     {
-        public void Exec(Eval.String msg)
+        public void Exec(String msg)
         {
             Debug.Log($"user msg: {msg}");
         }
