+++ Assets/Scripts/Interpreter/Eval/Type/Nil.cs
@@ -1,0 +1,16 @@
+namespace Interpreter.Eval.Type
+{
+    public class Nil : Value
+    {
+        public override string ToString()
+        {
+            return $"Nil";
+        }
+
+        public override bool Equals(Value other)
+        {
+            if (other == null) return false;
+            return GetType() == other.GetType();
+        }
+    }
+}
