+++ Assets/Scripts/Interpreter/Eval/Type/String.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: f8c0889ed76343598da250e082c5d2e6
+timeCreated: 1697857017
