--- Assets/Scripts/Interpreter/Eval/Value.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: fbc52a96f99248c091e80a748566d0ad
-timeCreated: 1697809395
